# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "E-Market",
     "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
     
       "version": "1.0"
   },
   "produces": ["application/json"],
   "paths": {
...
     "/api/mobile/profile/getProfile": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/profile/getProfile",
        "operationId": "index",
        "tags": ["Get Profile"],
        "description": "Get Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}

     

     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/profile/updateProfile": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/profile/updateProfile",
        "operationId": "index",
        "tags": ["Update Profile"],
        "description": "Update Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "firstName",  "in": "formData",  "required":true },
          { "name": "lastName",  "in": "formData"},
          { "name": "email",  "in": "formData",  "required":true },
          { "name": "profileImage",  "in": "formData",  "required":true ,"type":"file"},
          { "name": "address",  "in": "formData",  "required":true ,"type":"string"}
     

     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/cart/list": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/cart/list",
        "operationId": "index",
        "tags": ["Cart Listing"],
        "description": "get cart listing",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/cart/add": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/cart/add",
        "operationId": "index",
        "tags": ["Add to Cart"],
        "description": "Add Item to Cart",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "addressId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "serviceCharges",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "estimateDelivery",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "color",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "size",  "in": "formData" ,"type":"string" },
          { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"}



     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/cart/update": {
      "put": {
        "x-swagger-router-controller": "/api/mobile/cart/update",
        "operationId": "index",
        "tags": ["Update Cart"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "cartId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "addressId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "color",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "size",  "in": "formData" ,"type":"string" },
          { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"}
          

     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/cart/remove": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/cart/remove",
        "operationId": "index",
        "tags": ["Remove Item from Cart"],
        "description": "get Prduct listing by Category ID",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true"},
          { "name": "cartId",  "in": "query",  "required":"true"}


     ],

  
        "responses": {}
      }
    },
    "/api/mobile/cart/clear": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/cart/clear",
        "operationId": "index",
        "tags": ["Clear Cart"],
        "description": "get Prduct listing by Category ID",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true"}


     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/wishlist/list": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/wishlist/list",
        "operationId": "index",
        "tags": ["Wishlist Listing"],
        "description": "get wishlist listing",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "search",  "in": "query" ,"type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/wishlist/add": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/wishlist/add",
        "operationId": "index",
        "tags": ["Add to wishlist"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "serviceId",  "in": "formData",  "required":"true" ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":true ,"type":"string"}



     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/wishlist/remove": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/wishlist/remove",
        "operationId": "index",
        "tags": ["Remove from wishlist"],
        "description": "Remove from wishlist",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "wishId",  "in": "query",  "required":"true" ,"type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/wishlist/clearAll": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/wishlist/clearAll",
        "operationId": "index",
        "tags": ["Remove All from wishlist"],
        "description": "Remove all from wishlist",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}
     ],

   
    
    
        "responses": {}
      }
    },
...
   }
}